refactor: move play-by-plays to fixtures
