+++ packages/dota/src/__tests__/fixtures/gameEnd.ts
@@ -1,0 +1,1658 @@
+// disable prettier in this file
+
+// radiant win, invoker game
+export const gameEnd = ({
+  win_team,
+  matchId,
+  steam32,
+  steam64,
+  token,
+}: {
+  win_team: 'radiant' | 'dire'
+  matchId: string
+  steam32: string
+  steam64: string
+  token: string
+}) => [
+  {
+    events: [],
+    player: {},
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611094 },
+    auth: { token: token },
+  },
+  {
+    abilities: {},
+    buildings: { radiant: {} },
+    events: [],
+    hero: { id: -1 },
+    items: {},
+    map: {
+      name: 'start',
+      matchid: '0',
+      game_time: 5,
+      clock_time: 5,
+      daytime: false,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_INIT',
+      paused: false,
+      win_team: 'none',
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 0,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 0,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 600,
+      gold_reliable: 0,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 0,
+      gold_from_shared: 0,
+      gpm: 0,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611097 },
+    wearables: {},
+    added: {
+      player: {
+        steamid: true,
+        accountid: true,
+        name: true,
+        activity: true,
+        kills: true,
+        deaths: true,
+        assists: true,
+        last_hits: true,
+        denies: true,
+        kill_streak: true,
+        commands_issued: true,
+        kill_list: true,
+        team_name: true,
+        player_slot: true,
+        team_slot: true,
+        gold: true,
+        gold_reliable: true,
+        gold_unreliable: true,
+        gold_from_hero_kills: true,
+        gold_from_creep_kills: true,
+        gold_from_income: true,
+        gold_from_shared: true,
+        gpm: true,
+        xpm: true,
+      },
+    },
+    auth: { token: token },
+  },
+  {
+    abilities: {},
+    buildings: { radiant: {} },
+    events: [],
+    hero: { id: -1 },
+    items: {},
+    map: {
+      name: 'start',
+      matchid: '0',
+      game_time: 7,
+      clock_time: 7,
+      daytime: false,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_INIT',
+      paused: false,
+      win_team: 'none',
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 0,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 0,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 600,
+      gold_reliable: 0,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 0,
+      gold_from_shared: 0,
+      gpm: 0,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611099 },
+    wearables: {},
+    previously: { map: { game_time: 5, clock_time: 5 } },
+    auth: { token: token },
+  },
+  {
+    abilities: {},
+    buildings: { radiant: {} },
+    events: [],
+    hero: { id: -1 },
+    items: {},
+    map: {
+      name: 'start',
+      matchid: matchId,
+      game_time: 9,
+      clock_time: -74,
+      daytime: false,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_HERO_SELECTION',
+      paused: false,
+      win_team: 'none',
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 0,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 0,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 600,
+      gold_reliable: 0,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 0,
+      gold_from_shared: 0,
+      gpm: 0,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611101 },
+    wearables: {},
+    previously: { map: { game_time: 7, clock_time: 7, game_state: 'DOTA_GAMERULES_STATE_INIT' } },
+    auth: { token: token },
+  },
+  {
+    abilities: {},
+    buildings: { radiant: {} },
+    events: [],
+    hero: { id: 74, name: 'npc_dota_hero_invoker' },
+    items: {},
+    map: {
+      name: 'start',
+      matchid: matchId,
+      game_time: 10,
+      clock_time: -29,
+      daytime: false,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_STRATEGY_TIME',
+      paused: false,
+      win_team: 'none',
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 0,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 0,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 600,
+      gold_reliable: 0,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 0,
+      gold_from_shared: 0,
+      gpm: 0,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611105 },
+    wearables: {
+      wearable0: 5867,
+      wearable1: 23683,
+      wearable2: 98,
+      wearable3: 48,
+      wearable4: 4289,
+      wearable5: 8626,
+      wearable6: 6079,
+      wearable7: 8632,
+      wearable8: 13043,
+      wearable9: 13044,
+      wearable10: 13045,
+      wearable11: 13046,
+      wearable12: 13047,
+      wearable13: 13049,
+      wearable14: 13078,
+      wearable15: 766,
+      wearable25: 683,
+    },
+    previously: {
+      hero: { id: -1 },
+      map: { game_time: 9, clock_time: -74, game_state: 'DOTA_GAMERULES_STATE_HERO_SELECTION' },
+    },
+    added: {
+      hero: { name: true },
+      wearables: {
+        wearable0: true,
+        wearable1: true,
+        wearable2: true,
+        wearable3: true,
+        wearable4: true,
+        wearable5: true,
+        wearable6: true,
+        wearable7: true,
+        wearable8: true,
+        wearable9: true,
+        wearable10: true,
+        wearable11: true,
+        wearable12: true,
+        wearable13: true,
+        wearable14: true,
+        wearable15: true,
+        wearable25: true,
+      },
+    },
+    auth: { token: token },
+  },
+  {
+    abilities: {},
+    buildings: { radiant: {} },
+    events: [],
+    hero: { id: 74, name: 'npc_dota_hero_invoker' },
+    items: {},
+    map: {
+      name: 'start',
+      matchid: matchId,
+      game_time: 12,
+      clock_time: -27,
+      daytime: false,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_WAIT_FOR_MAP_TO_LOAD',
+      paused: false,
+      win_team: 'none',
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 0,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 0,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 600,
+      gold_reliable: 0,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 0,
+      gold_from_shared: 0,
+      gpm: 0,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611108 },
+    wearables: {
+      wearable0: 5867,
+      wearable1: 23683,
+      wearable2: 98,
+      wearable3: 48,
+      wearable4: 4289,
+      wearable5: 8626,
+      wearable6: 6079,
+      wearable7: 8632,
+      wearable8: 13043,
+      wearable9: 13044,
+      wearable10: 13045,
+      wearable11: 13046,
+      wearable12: 13047,
+      wearable13: 13049,
+      wearable14: 13078,
+      wearable15: 766,
+      wearable25: 683,
+    },
+    previously: {
+      map: { game_time: 10, clock_time: -29, game_state: 'DOTA_GAMERULES_STATE_STRATEGY_TIME' },
+    },
+    auth: { token: token },
+  },
+  {
+    abilities: {},
+    buildings: {
+      radiant: {
+        dota_goodguys_tower1_top: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_mid: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_bot: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower4_top: { health: 2600, max_health: 2600 },
+        dota_goodguys_tower4_bot: { health: 2600, max_health: 2600 },
+        good_rax_melee_top: { health: 2200, max_health: 2200 },
+        good_rax_range_top: { health: 1300, max_health: 1300 },
+        good_rax_melee_mid: { health: 2200, max_health: 2200 },
+        good_rax_range_mid: { health: 1300, max_health: 1300 },
+        good_rax_melee_bot: { health: 2200, max_health: 2200 },
+        good_rax_range_bot: { health: 1300, max_health: 1300 },
+        dota_goodguys_fort: { health: 4500, max_health: 4500 },
+      },
+    },
+    events: [],
+    hero: { id: 74, name: 'npc_dota_hero_invoker' },
+    items: {},
+    map: {
+      name: 'start',
+      matchid: matchId,
+      game_time: 14,
+      clock_time: -89,
+      daytime: false,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_PRE_GAME',
+      paused: false,
+      win_team: 'none',
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 0,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 0,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 600,
+      gold_reliable: 0,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 0,
+      gold_from_shared: 0,
+      gpm: 0,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611110 },
+    wearables: {
+      wearable0: 5867,
+      wearable1: 23683,
+      wearable2: 98,
+      wearable3: 48,
+      wearable4: 4289,
+      wearable5: 8626,
+      wearable6: 6079,
+      wearable7: 8632,
+      wearable8: 13043,
+      wearable9: 13044,
+      wearable10: 13045,
+      wearable11: 13046,
+      wearable12: 13047,
+      wearable13: 13049,
+      wearable14: 13078,
+      wearable15: 766,
+      wearable25: 683,
+    },
+    added: {
+      buildings: {
+        radiant: {
+          dota_goodguys_tower1_top: true,
+          dota_goodguys_tower2_top: true,
+          dota_goodguys_tower3_top: true,
+          dota_goodguys_tower1_mid: true,
+          dota_goodguys_tower2_mid: true,
+          dota_goodguys_tower3_mid: true,
+          dota_goodguys_tower1_bot: true,
+          dota_goodguys_tower2_bot: true,
+          dota_goodguys_tower3_bot: true,
+          dota_goodguys_tower4_top: true,
+          dota_goodguys_tower4_bot: true,
+          good_rax_melee_top: true,
+          good_rax_range_top: true,
+          good_rax_melee_mid: true,
+          good_rax_range_mid: true,
+          good_rax_melee_bot: true,
+          good_rax_range_bot: true,
+          dota_goodguys_fort: true,
+        },
+      },
+    },
+    previously: {
+      map: {
+        game_time: 12,
+        clock_time: -27,
+        game_state: 'DOTA_GAMERULES_STATE_WAIT_FOR_MAP_TO_LOAD',
+      },
+    },
+    auth: { token: token },
+  },
+  {
+    abilities: {
+      ability0: {
+        name: 'invoker_quas',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability1: {
+        name: 'invoker_wex',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability2: {
+        name: 'invoker_exort',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability3: {
+        name: 'invoker_empty1',
+        level: 0,
+        can_cast: false,
+        passive: true,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability4: {
+        name: 'invoker_empty2',
+        level: 0,
+        can_cast: false,
+        passive: true,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability5: {
+        name: 'invoker_invoke',
+        level: 1,
+        can_cast: true,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: true,
+      },
+    },
+    buildings: {
+      radiant: {
+        dota_goodguys_tower1_top: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_mid: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_bot: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower4_top: { health: 2600, max_health: 2600 },
+        dota_goodguys_tower4_bot: { health: 2600, max_health: 2600 },
+        good_rax_melee_top: { health: 2200, max_health: 2200 },
+        good_rax_range_top: { health: 1300, max_health: 1300 },
+        good_rax_melee_mid: { health: 2200, max_health: 2200 },
+        good_rax_range_mid: { health: 1300, max_health: 1300 },
+        good_rax_melee_bot: { health: 2200, max_health: 2200 },
+        good_rax_range_bot: { health: 1300, max_health: 1300 },
+        dota_goodguys_fort: { health: 4500, max_health: 4500 },
+      },
+    },
+    events: [],
+    hero: {
+      xpos: -6700,
+      ypos: -6700,
+      id: 74,
+      name: 'npc_dota_hero_invoker',
+      level: 1,
+      xp: 0,
+      alive: true,
+      respawn_seconds: 0,
+      buyback_cost: 253,
+      buyback_cooldown: 0,
+      health: 538,
+      max_health: 538,
+      health_percent: 100,
+      mana: 303,
+      max_mana: 303,
+      mana_percent: 100,
+      silenced: false,
+      stunned: false,
+      disarmed: false,
+      magicimmune: false,
+      hexed: false,
+      muted: false,
+      break: false,
+      aghanims_scepter: false,
+      aghanims_shard: false,
+      smoked: false,
+      has_debuff: false,
+      talent_1: false,
+      talent_2: false,
+      talent_3: false,
+      talent_4: false,
+      talent_5: false,
+      talent_6: false,
+      talent_7: false,
+      talent_8: false,
+      attributes_level: 0,
+    },
+    items: {
+      slot0: { name: 'empty' },
+      slot1: { name: 'empty' },
+      slot2: { name: 'empty' },
+      slot3: { name: 'empty' },
+      slot4: { name: 'empty' },
+      slot5: { name: 'empty' },
+      slot6: { name: 'empty' },
+      slot7: { name: 'empty' },
+      slot8: { name: 'empty' },
+      stash0: { name: 'empty' },
+      stash1: { name: 'empty' },
+      stash2: { name: 'empty' },
+      stash3: { name: 'empty' },
+      stash4: { name: 'empty' },
+      stash5: { name: 'empty' },
+      teleport0: {
+        name: 'item_tpscroll',
+        purchaser: 0,
+        item_level: 1,
+        can_cast: false,
+        cooldown: 99,
+        passive: false,
+        charges: 1,
+      },
+      neutral0: { name: 'empty' },
+    },
+    map: {
+      name: 'start',
+      matchid: matchId,
+      game_time: 16,
+      clock_time: -87,
+      daytime: false,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_PRE_GAME',
+      paused: false,
+      win_team: 'none',
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 0,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 0,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 600,
+      gold_reliable: 0,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 0,
+      gold_from_shared: 0,
+      gpm: 0,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611112 },
+    wearables: {
+      wearable0: 5867,
+      wearable1: 23683,
+      wearable2: 98,
+      wearable3: 48,
+      wearable4: 4289,
+      wearable5: 8626,
+      wearable6: 6079,
+      wearable7: 8632,
+      wearable8: 790,
+      wearable9: 791,
+      wearable10: 792,
+      wearable11: 793,
+      wearable12: 794,
+      wearable13: 795,
+      wearable14: 796,
+      wearable15: 797,
+      wearable16: 798,
+      wearable17: 683,
+      wearable18: 14912,
+    },
+    added: {
+      abilities: {
+        ability0: true,
+        ability1: true,
+        ability2: true,
+        ability3: true,
+        ability4: true,
+        ability5: true,
+      },
+      hero: {
+        xpos: true,
+        ypos: true,
+        level: true,
+        xp: true,
+        alive: true,
+        respawn_seconds: true,
+        buyback_cost: true,
+        buyback_cooldown: true,
+        health: true,
+        max_health: true,
+        health_percent: true,
+        mana: true,
+        max_mana: true,
+        mana_percent: true,
+        silenced: true,
+        stunned: true,
+        disarmed: true,
+        magicimmune: true,
+        hexed: true,
+        muted: true,
+        break: true,
+        aghanims_scepter: true,
+        aghanims_shard: true,
+        smoked: true,
+        has_debuff: true,
+        talent_1: true,
+        talent_2: true,
+        talent_3: true,
+        talent_4: true,
+        talent_5: true,
+        talent_6: true,
+        talent_7: true,
+        talent_8: true,
+        attributes_level: true,
+      },
+      items: {
+        slot0: true,
+        slot1: true,
+        slot2: true,
+        slot3: true,
+        slot4: true,
+        slot5: true,
+        slot6: true,
+        slot7: true,
+        slot8: true,
+        stash0: true,
+        stash1: true,
+        stash2: true,
+        stash3: true,
+        stash4: true,
+        stash5: true,
+        teleport0: true,
+        neutral0: true,
+      },
+      wearables: { wearable16: true, wearable17: true, wearable18: true },
+    },
+    previously: {
+      map: { game_time: 14, clock_time: -89 },
+      wearables: {
+        wearable8: 13043,
+        wearable9: 13044,
+        wearable10: 13045,
+        wearable11: 13046,
+        wearable12: 13047,
+        wearable13: 13049,
+        wearable14: 13078,
+        wearable15: 766,
+        wearable25: 683,
+      },
+    },
+    auth: { token: token },
+  },
+  {
+    abilities: {
+      ability0: {
+        name: 'invoker_quas',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability1: {
+        name: 'invoker_wex',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability2: {
+        name: 'invoker_exort',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability3: {
+        name: 'invoker_empty1',
+        level: 0,
+        can_cast: false,
+        passive: true,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability4: {
+        name: 'invoker_empty2',
+        level: 0,
+        can_cast: false,
+        passive: true,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability5: {
+        name: 'invoker_invoke',
+        level: 1,
+        can_cast: true,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: true,
+      },
+    },
+    buildings: {
+      radiant: {
+        dota_goodguys_tower1_top: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_mid: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_bot: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower4_top: { health: 2600, max_health: 2600 },
+        dota_goodguys_tower4_bot: { health: 2600, max_health: 2600 },
+        good_rax_melee_top: { health: 2200, max_health: 2200 },
+        good_rax_range_top: { health: 1300, max_health: 1300 },
+        good_rax_melee_mid: { health: 2200, max_health: 2200 },
+        good_rax_range_mid: { health: 1300, max_health: 1300 },
+        good_rax_melee_bot: { health: 2200, max_health: 2200 },
+        good_rax_range_bot: { health: 1300, max_health: 1300 },
+        dota_goodguys_fort: { health: 4500, max_health: 4500 },
+      },
+    },
+    events: [],
+    hero: {
+      xpos: -6700,
+      ypos: -6700,
+      id: 74,
+      name: 'npc_dota_hero_invoker',
+      level: 1,
+      xp: 0,
+      alive: true,
+      respawn_seconds: 0,
+      buyback_cost: 253,
+      buyback_cooldown: 0,
+      health: 538,
+      max_health: 538,
+      health_percent: 100,
+      mana: 303,
+      max_mana: 303,
+      mana_percent: 100,
+      silenced: false,
+      stunned: false,
+      disarmed: false,
+      magicimmune: false,
+      hexed: false,
+      muted: false,
+      break: false,
+      aghanims_scepter: false,
+      aghanims_shard: false,
+      smoked: false,
+      has_debuff: false,
+      talent_1: false,
+      talent_2: false,
+      talent_3: false,
+      talent_4: false,
+      talent_5: false,
+      talent_6: false,
+      talent_7: false,
+      talent_8: false,
+      attributes_level: 0,
+    },
+    items: {
+      slot0: { name: 'empty' },
+      slot1: { name: 'empty' },
+      slot2: { name: 'empty' },
+      slot3: { name: 'empty' },
+      slot4: { name: 'empty' },
+      slot5: { name: 'empty' },
+      slot6: { name: 'empty' },
+      slot7: { name: 'empty' },
+      slot8: { name: 'empty' },
+      stash0: { name: 'empty' },
+      stash1: { name: 'empty' },
+      stash2: { name: 'empty' },
+      stash3: { name: 'empty' },
+      stash4: { name: 'empty' },
+      stash5: { name: 'empty' },
+      teleport0: {
+        name: 'item_tpscroll',
+        purchaser: 0,
+        item_level: 1,
+        can_cast: false,
+        cooldown: 97,
+        passive: false,
+        charges: 1,
+      },
+      neutral0: { name: 'empty' },
+    },
+    map: {
+      name: 'start',
+      matchid: matchId,
+      game_time: 17,
+      clock_time: -86,
+      daytime: false,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_PRE_GAME',
+      paused: false,
+      win_team: 'none',
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 0,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 0,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 600,
+      gold_reliable: 0,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 0,
+      gold_from_shared: 0,
+      gpm: 0,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611114 },
+    wearables: {
+      wearable0: 5867,
+      wearable1: 23683,
+      wearable2: 98,
+      wearable3: 48,
+      wearable4: 4289,
+      wearable5: 8626,
+      wearable6: 6079,
+      wearable7: 8632,
+      wearable8: 790,
+      wearable9: 791,
+      wearable10: 792,
+      wearable11: 793,
+      wearable12: 794,
+      wearable13: 795,
+      wearable14: 796,
+      wearable15: 797,
+      wearable16: 798,
+      wearable17: 683,
+      wearable18: 14912,
+    },
+    previously: { items: { teleport0: { cooldown: 99 } }, map: { game_time: 16, clock_time: -87 } },
+    auth: { token: token },
+  },
+  {
+    abilities: {
+      ability0: {
+        name: 'invoker_quas',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability1: {
+        name: 'invoker_wex',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability2: {
+        name: 'invoker_exort',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability3: {
+        name: 'invoker_empty1',
+        level: 0,
+        can_cast: false,
+        passive: true,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability4: {
+        name: 'invoker_empty2',
+        level: 0,
+        can_cast: false,
+        passive: true,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability5: {
+        name: 'invoker_invoke',
+        level: 1,
+        can_cast: true,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: true,
+      },
+    },
+    buildings: {
+      radiant: {
+        dota_goodguys_tower1_top: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_mid: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_bot: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower4_top: { health: 2600, max_health: 2600 },
+        dota_goodguys_tower4_bot: { health: 2600, max_health: 2600 },
+        good_rax_melee_top: { health: 2200, max_health: 2200 },
+        good_rax_range_top: { health: 1300, max_health: 1300 },
+        good_rax_melee_mid: { health: 2200, max_health: 2200 },
+        good_rax_range_mid: { health: 1300, max_health: 1300 },
+        good_rax_melee_bot: { health: 2200, max_health: 2200 },
+        good_rax_range_bot: { health: 1300, max_health: 1300 },
+        dota_goodguys_fort: { health: 4500, max_health: 4500 },
+      },
+    },
+    events: [],
+    hero: {
+      xpos: -6834,
+      ypos: -6473,
+      id: 74,
+      name: 'npc_dota_hero_invoker',
+      level: 1,
+      xp: 0,
+      alive: true,
+      respawn_seconds: 0,
+      buyback_cost: 253,
+      buyback_cooldown: 0,
+      health: 538,
+      max_health: 538,
+      health_percent: 100,
+      mana: 303,
+      max_mana: 303,
+      mana_percent: 100,
+      silenced: false,
+      stunned: false,
+      disarmed: false,
+      magicimmune: false,
+      hexed: false,
+      muted: false,
+      break: false,
+      aghanims_scepter: false,
+      aghanims_shard: false,
+      smoked: false,
+      has_debuff: false,
+      talent_1: false,
+      talent_2: false,
+      talent_3: false,
+      talent_4: false,
+      talent_5: false,
+      talent_6: false,
+      talent_7: false,
+      talent_8: false,
+      attributes_level: 0,
+    },
+    items: {
+      slot0: { name: 'empty' },
+      slot1: { name: 'empty' },
+      slot2: { name: 'empty' },
+      slot3: { name: 'empty' },
+      slot4: { name: 'empty' },
+      slot5: { name: 'empty' },
+      slot6: { name: 'empty' },
+      slot7: { name: 'empty' },
+      slot8: { name: 'empty' },
+      stash0: { name: 'empty' },
+      stash1: { name: 'empty' },
+      stash2: { name: 'empty' },
+      stash3: { name: 'empty' },
+      stash4: { name: 'empty' },
+      stash5: { name: 'empty' },
+      teleport0: {
+        name: 'item_tpscroll',
+        purchaser: 0,
+        item_level: 1,
+        can_cast: false,
+        cooldown: 96,
+        passive: false,
+        charges: 1,
+      },
+      neutral0: { name: 'empty' },
+    },
+    map: {
+      name: 'start',
+      matchid: matchId,
+      game_time: 19,
+      clock_time: -84,
+      daytime: false,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_PRE_GAME',
+      paused: false,
+      win_team: 'none',
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 0,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 0,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 600,
+      gold_reliable: 0,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 0,
+      gold_from_shared: 0,
+      gpm: 0,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611116 },
+    wearables: {
+      wearable0: 5867,
+      wearable1: 23683,
+      wearable2: 98,
+      wearable3: 48,
+      wearable4: 4289,
+      wearable5: 8626,
+      wearable6: 6079,
+      wearable7: 8632,
+      wearable8: 790,
+      wearable9: 791,
+      wearable10: 792,
+      wearable11: 793,
+      wearable12: 794,
+      wearable13: 795,
+      wearable14: 796,
+      wearable15: 797,
+      wearable16: 798,
+      wearable17: 683,
+      wearable18: 14912,
+    },
+    previously: {
+      hero: { xpos: -6700, ypos: -6700 },
+      items: { teleport0: { cooldown: 97 } },
+      map: { game_time: 17, clock_time: -86 },
+    },
+    auth: { token: token },
+  },
+  {
+    abilities: {
+      ability0: {
+        name: 'invoker_quas',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability1: {
+        name: 'invoker_wex',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability2: {
+        name: 'invoker_exort',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability3: {
+        name: 'invoker_empty1',
+        level: 0,
+        can_cast: false,
+        passive: true,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability4: {
+        name: 'invoker_empty2',
+        level: 0,
+        can_cast: false,
+        passive: true,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability5: {
+        name: 'invoker_invoke',
+        level: 1,
+        can_cast: true,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: true,
+      },
+    },
+    buildings: {
+      radiant: {
+        dota_goodguys_tower1_top: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_mid: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_bot: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower4_top: { health: 2600, max_health: 2600 },
+        dota_goodguys_tower4_bot: { health: 2600, max_health: 2600 },
+        good_rax_melee_top: { health: 2200, max_health: 2200 },
+        good_rax_range_top: { health: 1300, max_health: 1300 },
+        good_rax_melee_mid: { health: 2200, max_health: 2200 },
+        good_rax_range_mid: { health: 1300, max_health: 1300 },
+        good_rax_melee_bot: { health: 2200, max_health: 2200 },
+        good_rax_range_bot: { health: 1300, max_health: 1300 },
+        dota_goodguys_fort: { health: 4500, max_health: 4500 },
+      },
+    },
+    events: [],
+    hero: {
+      xpos: -6878,
+      ypos: -6400,
+      id: 74,
+      name: 'npc_dota_hero_invoker',
+      level: 1,
+      xp: 0,
+      alive: true,
+      respawn_seconds: 0,
+      buyback_cost: 254,
+      buyback_cooldown: 0,
+      health: 538,
+      max_health: 538,
+      health_percent: 100,
+      mana: 303,
+      max_mana: 303,
+      mana_percent: 100,
+      silenced: false,
+      stunned: false,
+      disarmed: false,
+      magicimmune: false,
+      hexed: false,
+      muted: false,
+      break: false,
+      aghanims_scepter: false,
+      aghanims_shard: false,
+      smoked: false,
+      has_debuff: false,
+      talent_1: false,
+      talent_2: false,
+      talent_3: false,
+      talent_4: false,
+      talent_5: false,
+      talent_6: false,
+      talent_7: false,
+      talent_8: false,
+      attributes_level: 0,
+    },
+    items: {
+      slot0: { name: 'empty' },
+      slot1: { name: 'empty' },
+      slot2: { name: 'empty' },
+      slot3: { name: 'empty' },
+      slot4: { name: 'empty' },
+      slot5: { name: 'empty' },
+      slot6: { name: 'empty' },
+      slot7: { name: 'empty' },
+      slot8: { name: 'empty' },
+      stash0: { name: 'empty' },
+      stash1: { name: 'empty' },
+      stash2: { name: 'empty' },
+      stash3: { name: 'empty' },
+      stash4: { name: 'empty' },
+      stash5: { name: 'empty' },
+      teleport0: {
+        name: 'item_tpscroll',
+        purchaser: 0,
+        item_level: 1,
+        can_cast: false,
+        cooldown: 94,
+        passive: false,
+        charges: 1,
+      },
+      neutral0: { name: 'empty' },
+    },
+    map: {
+      name: 'start',
+      matchid: matchId,
+      game_time: 21,
+      clock_time: 0,
+      daytime: true,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_GAME_IN_PROGRESS',
+      paused: false,
+      win_team: 'none',
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 0,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 0,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 602,
+      gold_reliable: 2,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 2,
+      gold_from_shared: 0,
+      gpm: 174,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611117 },
+    wearables: {
+      wearable0: 5867,
+      wearable1: 23683,
+      wearable2: 98,
+      wearable3: 48,
+      wearable4: 4289,
+      wearable5: 8626,
+      wearable6: 6079,
+      wearable7: 8632,
+      wearable8: 790,
+      wearable9: 791,
+      wearable10: 792,
+      wearable11: 793,
+      wearable12: 794,
+      wearable13: 795,
+      wearable14: 796,
+      wearable15: 797,
+      wearable16: 798,
+      wearable17: 683,
+      wearable18: 14912,
+    },
+    previously: {
+      hero: { xpos: -6834, ypos: -6473, buyback_cost: 253 },
+      items: { teleport0: { cooldown: 96 } },
+      map: {
+        game_time: 19,
+        clock_time: -84,
+        daytime: false,
+        game_state: 'DOTA_GAMERULES_STATE_PRE_GAME',
+      },
+      player: { gold: 600, gold_reliable: 0, gold_from_income: 0, gpm: 0 },
+    },
+    auth: { token: token },
+  },
+  {
+    abilities: {
+      ability0: {
+        name: 'invoker_quas',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability1: {
+        name: 'invoker_wex',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability2: {
+        name: 'invoker_exort',
+        level: 0,
+        can_cast: false,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability3: {
+        name: 'invoker_empty1',
+        level: 0,
+        can_cast: false,
+        passive: true,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability4: {
+        name: 'invoker_empty2',
+        level: 0,
+        can_cast: false,
+        passive: true,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: false,
+      },
+      ability5: {
+        name: 'invoker_invoke',
+        level: 1,
+        can_cast: true,
+        passive: false,
+        ability_active: true,
+        cooldown: 0,
+        ultimate: true,
+      },
+    },
+    buildings: {
+      radiant: {
+        dota_goodguys_tower1_top: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_top: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_mid: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_mid: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower1_bot: { health: 1800, max_health: 1800 },
+        dota_goodguys_tower2_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower3_bot: { health: 2500, max_health: 2500 },
+        dota_goodguys_tower4_top: { health: 2600, max_health: 2600 },
+        dota_goodguys_tower4_bot: { health: 2600, max_health: 2600 },
+        good_rax_melee_top: { health: 2200, max_health: 2200 },
+        good_rax_range_top: { health: 1300, max_health: 1300 },
+        good_rax_melee_mid: { health: 2200, max_health: 2200 },
+        good_rax_range_mid: { health: 1300, max_health: 1300 },
+        good_rax_melee_bot: { health: 2200, max_health: 2200 },
+        good_rax_range_bot: { health: 1300, max_health: 1300 },
+        dota_goodguys_fort: { health: 4500, max_health: 4500 },
+      },
+    },
+    events: [],
+    hero: {
+      xpos: -6878,
+      ypos: -6400,
+      id: 74,
+      name: 'npc_dota_hero_invoker',
+      level: 1,
+      xp: 0,
+      alive: true,
+      respawn_seconds: 0,
+      buyback_cost: 254,
+      buyback_cooldown: 0,
+      health: 538,
+      max_health: 538,
+      health_percent: 100,
+      mana: 303,
+      max_mana: 303,
+      mana_percent: 100,
+      silenced: false,
+      stunned: false,
+      disarmed: false,
+      magicimmune: false,
+      hexed: false,
+      muted: false,
+      break: false,
+      aghanims_scepter: false,
+      aghanims_shard: false,
+      smoked: false,
+      has_debuff: false,
+      talent_1: false,
+      talent_2: false,
+      talent_3: false,
+      talent_4: false,
+      talent_5: false,
+      talent_6: false,
+      talent_7: false,
+      talent_8: false,
+      attributes_level: 0,
+    },
+    items: {
+      slot0: { name: 'empty' },
+      slot1: { name: 'empty' },
+      slot2: { name: 'empty' },
+      slot3: { name: 'empty' },
+      slot4: { name: 'empty' },
+      slot5: { name: 'empty' },
+      slot6: { name: 'empty' },
+      slot7: { name: 'empty' },
+      slot8: { name: 'empty' },
+      stash0: { name: 'empty' },
+      stash1: { name: 'empty' },
+      stash2: { name: 'empty' },
+      stash3: { name: 'empty' },
+      stash4: { name: 'empty' },
+      stash5: { name: 'empty' },
+      teleport0: {
+        name: 'item_tpscroll',
+        purchaser: 0,
+        item_level: 1,
+        can_cast: false,
+        cooldown: 92,
+        passive: false,
+        charges: 1,
+      },
+      neutral0: { name: 'empty' },
+    },
+    map: {
+      name: 'start',
+      matchid: matchId,
+      game_time: 23,
+      clock_time: 1,
+      daytime: true,
+      nightstalker_night: false,
+      radiant_score: 0,
+      dire_score: 0,
+      game_state: 'DOTA_GAMERULES_STATE_POST_GAME',
+      paused: false,
+      win_team: win_team,
+      customgamename: '',
+      ward_purchase_cooldown: 0,
+    },
+    player: {
+      steamid: steam64,
+      accountid: steam32,
+      name: 'loremipsum',
+      activity: 'playing',
+      kills: 0,
+      deaths: 0,
+      assists: 0,
+      last_hits: 1,
+      denies: 0,
+      kill_streak: 0,
+      commands_issued: 2,
+      kill_list: {},
+      team_name: 'radiant',
+      player_slot: 0,
+      team_slot: 0,
+      gold: 603,
+      gold_reliable: 3,
+      gold_unreliable: 600,
+      gold_from_hero_kills: 0,
+      gold_from_creep_kills: 0,
+      gold_from_income: 3,
+      gold_from_shared: 0,
+      gpm: 125,
+      xpm: 0,
+    },
+    provider: { name: 'Dota 2', appid: 570, version: 47, timestamp: 1693611119 },
+    wearables: {
+      wearable0: 5867,
+      wearable1: 23683,
+      wearable2: 98,
+      wearable3: 48,
+      wearable4: 4289,
+      wearable5: 8626,
+      wearable6: 6079,
+      wearable7: 8632,
+      wearable8: 790,
+      wearable9: 791,
+      wearable10: 792,
+      wearable11: 793,
+      wearable12: 794,
+      wearable13: 795,
+      wearable14: 796,
+      wearable15: 797,
+      wearable16: 798,
+      wearable17: 683,
+      wearable18: 14912,
+    },
+    previously: {
+      items: { teleport0: { cooldown: 94 } },
+      map: {
+        game_time: 21,
+        clock_time: 0,
+        game_state: 'DOTA_GAMERULES_STATE_GAME_IN_PROGRESS',
+        win_team: 'none',
+      },
+      player: {
+        last_hits: 0,
+        commands_issued: 0,
+        gold: 602,
+        gold_reliable: 2,
+        gold_from_income: 2,
+        gpm: 174,
+      },
+    },
+    auth: { token: token },
+  },
+]
